+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 based on [Semantic Versioning](http://semver.org/)
 
-*2.0.0* 2014-03-29
+*2.0.0* 2014-03-31
 
++ add a list of not increment attributes in concept
 + add a list of identifier attributes in concept
-+ add a list of non identifier attributes in concept
++ add a list of not identifier attributes in concept
 + add missing end of a generated test
 + validate if code is empty in lower or upper transformations in entity
 + add error messages when add and remove propagations are not successful